automatically backed up by learn

+++ index.js
@@ -27,5 +27,5 @@ function makeInt(n) {
 }
 
 function makeInt(n) {
-  parseInt(n, 10)
+  parseFloat(n, 10)
 }
